chore: tested exercise

+++ src/shapes/Measurable.java
@@ -1,8 +1,9 @@
 package shapes;
 
 public interface Measurable {
-    double getPerimeter();
-    double getArea();
+    int getPerimeter();
+    int getArea();
 
-    
+
+
 }
